# apps/server/swagger.yaml
openapi: 3.0.3
info:
  title: Think Different Academy – Tour de App 2026 API
  version: 1.0.0
  description: >
    Official OpenAPI (Swagger) specification for the **Tour de App 2026** competition project.

    This API defines the architecture for the **Think Different Academy** learning platform —
    an interactive educational system combining courses, materials, quizzes, and live updates.

    It covers the main development phases of the challenge:
    - Phase 0 – basic deployment (Hello TdA)
    - Phase 1 – course catalog and instructor dashboard
    - Phase 2 – course materials
    - Phase 3 – quizzes
    - Phase 4 – news feed and SSE updates

    **Authentication is not part of this specification.**
    All endpoints are publicly accessible for simplicity.
servers:
  - url: https://<identificator>.tourde.app/api
    description: Production deployment environment for Tour de App submissions
  - url: https://uwu-t34m-su8d0m41n.tourde.app/api
    description: Example test environment for a demo team
paths:
...
  /courses/{courseId}/quizzes:
    parameters:
      - $ref: '#/components/parameters/CourseId'
    get:
      summary: List quizzes
      description: Returns all quizzes belonging to a specific course.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Quiz'
    post:
      summary: Create new quiz
      description: Instructor can create a new quiz for a course.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Quiz'
      responses:
        '201':
          description: Quiz created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quiz'
  /courses/{courseId}/quizzes/{quizId}:
    parameters:
      - $ref: '#/components/parameters/CourseId'
      - $ref: '#/components/parameters/QuizId'
    get:
      summary: Get quiz details
      description: Returns quiz data, including all questions.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quiz'
    put:
      summary: Update quiz
      description: Update quiz title or questions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Quiz'
      responses:
        '200':
          description: Quiz updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quiz'
    delete:
      summary: Delete quiz
      description: Permanently removes the quiz from the course.
      responses:
        '204':
          description: Quiz deleted
...
components:
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    QuizId:
      name: quizId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Quiz:
      type: object
      properties:
        uuid:
          type: string
        title:
          type: string
        attemptsCount:
          type: integer
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
      required: [title, questions]
    Question:
      oneOf:
        - $ref: '#/components/schemas/SingleChoiceQuestion'
        - $ref: '#/components/schemas/MultipleChoiceQuestion'
      discriminator:
        propertyName: type
    SingleChoiceQuestion:
      type: object
      properties:
        uuid:
          type: string
        type:
          type: string
          enum: [singleChoice]
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctIndex:
          type: integer
      required: [type, question, options, correctIndex]
    MultipleChoiceQuestion:
      type: object
      properties:
        uuid:
          type: string
        type:
          type: string
          enum: [multipleChoice]
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctIndices:
          type: array
          items:
            type: integer
      required: [type, question, options, correctIndices]
...